requirements.txt
customtkinter==5.2.2
from_root==1.3.0
numpy==1.23.5
paramiko==2.6.0
Pillow==7.0.0
PyYAML==6.0.1
setuptools==70.2.0
typing_extensions==4.12.2
psutil==6.1.1